Factor repositories from services
Distinguish between transformations and parameters that are generic to a
service and those that are specific to particular instance of that
service for a dataset, called repositories. This both makes these
distinctions more explicit and clear (e.g., the case where there was a
rotation between the export of the Hemibrain N5 and the Neuprint
coordinate system), and make the services themselves reusable for other
datasets or instances.

This should also restore the ability to predict NTs for datasets without
specifying or applying transforms from any repository or service.